hide teh picto functionality is working as intended
