+++ lib/presentation/screens/create_picto/choose_picto_day_screen.dart
@@ -3,6 +3,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:ottaa_project_flutter/application/common/extensions/translate_string.dart';
 import 'package:ottaa_project_flutter/application/providers/create_picto_provider.dart';
 import 'package:ottaa_project_flutter/presentation/common/widgets/simple_button.dart';
+import 'package:ottaa_project_flutter/presentation/screens/create_picto/ui/day_widget.dart';
+import 'package:ottaa_project_flutter/presentation/screens/create_picto/ui/time_widget.dart';
 import 'package:ottaa_project_flutter/presentation/screens/create_picto/widgets/image_widget.dart';
 
 class ChoosePictoDayScreen extends ConsumerWidget {
@@ -93,80 +95,6 @@ class ChoosePictoDayScreen extends ConsumerWidget {
             ),
           ),
         ],
-      ),
-    );
-  }
-}
-
-class DayWidget extends ConsumerWidget {
-  const DayWidget({
-    super.key,
-    required this.text,
-  });
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final provider = ref.watch(createPictoProvider);
-    final textTheme = Theme.of(context).textTheme;
-    final colorScheme = Theme.of(context).colorScheme;
-    return GestureDetector(
-      onTap: () {
-        if (provider.daysToUsePicto.contains(text)) {
-          provider.daysToUsePicto.remove(text);
-        } else {
-          provider.daysToUsePicto.add(text);
-        }
-        provider.notify();
-      },
-      child: Container(
-        decoration: BoxDecoration(
-          color: provider.daysToUsePicto.contains(text) ? colorScheme.primary : Colors.white,
-          borderRadius: BorderRadius.circular(8),
-        ),
-        padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-        child: Text(
-          text,
-          style: textTheme.bodySmall!.copyWith(color: provider.daysToUsePicto.contains(text) ? Colors.white : Colors.grey),
-        ),
-      ),
-    );
-  }
-}
-
-class TimeWidget extends ConsumerWidget {
-  const TimeWidget({
-    super.key,
-    required this.text,
-  });
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final provider = ref.watch(createPictoProvider);
-    final textTheme = Theme.of(context).textTheme;
-    final colorScheme = Theme.of(context).colorScheme;
-    return GestureDetector(
-      onTap: () {
-        if (provider.timeForPicto.contains(text)) {
-          provider.timeForPicto.remove(text);
-        } else {
-          provider.timeForPicto.add(text);
-        }
-        provider.notify();
-      },
-      child: Container(
-        decoration: BoxDecoration(
-          color: provider.timeForPicto.contains(text) ? colorScheme.primary : Colors.white,
-          borderRadius: BorderRadius.circular(8),
-        ),
-        padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-        child: Text(
-          text,
-          style: textTheme.bodySmall!.copyWith(color: provider.timeForPicto.contains(text) ? Colors.white : Colors.grey),
-        ),
       ),
     );
   }
